Define calendar objects

Generate a calendar object of class "calcal". Examples of calendars produced in this way include cal_gregorian, cal_hebrew, cal_islamic, cal_iso, cal_julian, and cal_roman.

Usage

cal_calendar(
  name,
  short_name,
  epoch,
  granularities,
  check_granularities,
  format,
  from_rd,
  to_rd
)

Arguments

name

Name of calendar

short_name

Short name of calendar

epoch

Epoch of calendar

granularities

Granularities of calendar (e.g., for the Gregorian calendar, the granularities are year, month, and day).

check_granularities

Function to check granularities are valid (e.g., Gregorian months should be between 1 and 12).

format

Functon to specify date format as a character string.

from_rd

Function to convert from RD to calendar date.

to_rd

Function to convert from calendar date to RD.