ISO calendar dates

Work with ISO 8601 date objects. Weeks are defined as starting on Mondays. Week 1 is the first week with at least 4 days in the year. Equivalently, it is the week containing 4 January. There is no week 0; instead week 1 of a year may begin in the previous calendar year.

Usage

cal_iso

Format

An object of class calcal of length 8.

Details

More flexible week numbering is possible using Gregorian dates with week_of_year().

See also

Examples

iso <- new_date(year = 2025, week = 23, day = 2:4, calendar = cal_iso)
iso
#> <iso[3]>
#> [1] 2025-23-02 2025-23-03 2025-23-04
as_date(iso, calendar = cal_gregorian)
#> <gregorian[3]>
#> [1] 2025-Jun-03 2025-Jun-04 2025-Jun-05
greg <- new_date(year = 2025, month = 1, day = 1:4, calendar = cal_gregorian)
as_date(greg, calendar = cal_iso)
#> <iso[4]>
#> [1] 2025-01-03 2025-01-04 2025-01-05 2025-01-06