Linear Granularity

Linear granularities are produced by combining the year and a cyclic granularity.

Usage

linear_granularity(gran_fun, gran_levels, gran_name)

Arguments

gran_fun

A function that extracts a cyclic granularity from a date variable. This function must return an integer.

gran_levels

A character vector of the granularity levels.

gran_name

A character string giving the name of the granularity

Value

A function that creates a linear granularity object.

Examples

# Compute quarter from a Gregorian date
quarter <- function(date) {
  ceiling(month_of_year(date) / 3)
}
year_quarter <- linear_granularity(quarter, paste0("Q",1:4), "year-quarter")
year_quarter(gregorian_date(2025, 6, 25) + 1:10)
#> Error in factor(gran, labels = gran_levels, ordered = TRUE): invalid 'labels'; length 4 should be 1 or 2