Baháʼí dates

The Baháʼí calendar is a solar calendar used in the Baháʼí faith comprising 18 months, with four or five intercalary days each year. The New Year is at the northern Spring equinox, corresponding to 21 March on the Gregorian calendar.

Usage

bahai_date(major, cycle, year, month, day)

as_bahai(date)

Arguments

major

A numeric vector of major periods

cycle

A numeric vector of cycles

year

A numeric vector of years within the cycles

month

A numeric vector of months

day

A numeric vector of days

date

A numeric vector of dates

Value

A bahai vector object

See also

Examples

tibble::tibble(
 gregorian = gregorian_date(2025, 1, 1:31),
 bahai = as_bahai(gregorian)
)
#> # A tibble: 31 × 2
#>     gregorian          bahai
#>         <Gre>          <Bah>
#>  1 2025-NA-01 01-10-10-16-02
#>  2 2025-NA-02 01-10-10-16-03
#>  3 2025-NA-03 01-10-10-16-04
#>  4 2025-NA-04 01-10-10-16-05
#>  5 2025-NA-05 01-10-10-16-06
#>  6 2025-NA-06 01-10-10-16-07
#>  7 2025-NA-07 01-10-10-16-08
#>  8 2025-NA-08 01-10-10-16-09
#>  9 2025-NA-09 01-10-10-16-10
#> 10 2025-NA-10 01-10-10-16-11
#> # ℹ 21 more rows
bahai_date(1, 10, 11, 3, 5)
#> <bahai[1]>
#> [1] 01-10-11-3-05
as_bahai("2016-01-01")
#> <bahai[1]>
#> [1] 01-10-01-16-02
as_bahai(Sys.Date())
#> <bahai[1]>
#> [1] 01-10-11-5-11